test(integration): one spark-parity file per RS_ function
Move the RS_GeoTransform cases into their own test_rs_geotransform.py,
rename test_rs_scalar.py -> test_rs_bandnodatavalue.py and
test_rs_raster_out.py -> test_rs_setbandnodatavalue.py to match their
single-function contents, and record the file-per-function convention in
the suite README.
