+++ integration/spark-parity/README.md
@@ -65,6 +65,9 @@ Useful knobs, both read by `sedonadb.testing_spark`:
 
 ## Conventions
 
+One test file per RS_ function, named `test_rs_<function>.py` — parity coverage
+for a function lands in its own file, not appended to a neighbor's.
+
 Where the two engines are known to diverge and we intend to close the gap, mark
 the case `xfail(reason=...)` rather than deleting or loosening it. The suite then
 doubles as a catalog of what to fix, and flips to `xpass` the day the fix lands.
